extract chat component
